updated USAGE.md

+++ src/actions/removeProj.ts
@@ -7,22 +7,26 @@ const STORAGE_PATH = homedir() + "/.tasklist/tasklist.json";
 
 let tasks = JSON.parse(fs.readFileSync(STORAGE_PATH, "utf-8"));
 
-// function to list pending tasks.
-export default async function delProj(proj: string) {
-  delete tasks[proj];
-  console.clear();
-  const confirmDelete = await inquirer.prompt({
-    name: "choice",
-    type: "confirm",
-    message: `Are you sure to delete ${proj}?`,
-    default() {
-      return false;
-    },
-  });
-  if (confirmDelete.choice) {
-    console.log(
-      `Project ${chalk.yellow(proj)} has been ${chalk.red("deleted.")}`
-    );
-    fs.writeFileSync(STORAGE_PATH, JSON.stringify(tasks));
+// function to remove a project
+export default async function removeProj(proj: string) {
+  if (Object.keys(tasks).includes(proj)) {
+    delete tasks[proj];
+    console.clear();
+    const confirmDelete = await inquirer.prompt({
+      name: "choice",
+      type: "confirm",
+      message: `Are you sure to delete ${proj}?`,
+      default() {
+        return false;
+      },
+    });
+    if (confirmDelete.choice) {
+      console.log(
+        `Project ${chalk.yellow(proj)} has been ${chalk.red("deleted.")}`
+      );
+      fs.writeFileSync(STORAGE_PATH, JSON.stringify(tasks));
+    }
+  } else {
+    console.log(`${chalk.red(proj)}: No such project.`);
   }
 }
